rototill and fix build
Disable perl and java too, because the build was trying to use
them and was trying to install into into someplace it ought not.

This is an embedded use; disable the tests on the assumption that
we're using a well tested use point release, and we don't run the
tests anyway.

With all of this flag twiddling going on, sort the flags to show
what is being enabled and what is being disabled.